VmCpu: Move PMU state to its own module
It's pretty self-contained and can be easily moved into a separate
module. Also fix some overly-long comment lines while we're at it.
